requirements.txt
streamlit==1.50.0
supabase==2.28.3
braintree==4.42.0
boto3==1.42.76
python-dotenv==1.2.1
requests==2.32.5
pandas==2.3.3
